ai-models/requirements.txt
# AI Models Requirements

# Deep Learning Framework
torch>=2.3.0,<2.4.0
torchvision>=0.18.0,<0.19.0
torchaudio>=2.3.0,<2.4.0

# ONNX for model optimization
onnx>=1.16.0,<1.17.0
onnxruntime>=1.18.0,<1.19.0

# Computer Vision
opencv-python>=4.9.0,<5.0.0
Pillow>=10.3.0,<11.0.0

# Audio Processing
librosa>=0.10.0,<0.11.0
soundfile>=0.12.0,<0.13.0
pydub>=0.25.0,<0.26.0

# Video Processing
ffmpeg-python>=0.2.0,<0.3.0

# Scientific Computing
numpy>=1.26.0,<2.0.0
scipy>=1.13.0,<1.14.0

# Utilities
tqdm>=4.66.0,<4.67.0
pyyaml>=6.0.0,<7.0.0

# Training utilities
tensorboard>=2.16.0,<2.17.0
wandb>=0.17.0,<0.18.0

# Testing
pytest>=8.2.0,<8.3.0
